chore: cleaned up admission test
